moved the vacancy-details page to the prototype folder, made a simple career-ru.json file
