--- src/atomic/page/vacancy-details.tsx
@@ -1,365 +1,0 @@
-import React from 'react';
-import NavigationBar from "@/atomic/organism/navbar";
-import BulletList from '../molecule/bullet-list';
-
-interface VacancyDetailsProps {
-    position: string;
-}
-
-const vacancyData: { [key: string]: any } = {
-    webdesigner: {
-        title: 'Web Designer🎨',
-        description: 
-        `Мы ищем стажера веб-дизайнера, который поможет создавать современные и привлекательные дизайны для 
-        нашего сайта и мобильных приложений. Вы будете работать с командой разработчиков, 
-        чтобы воплотить ваши идеи в жизнь.`,
-        expectations: [
-            { 
-                title: 'Дизайн сайта', 
-                items: [ 
-                    'Опыт работы с основными графическими редакторами',
-                    'Знание принципов UI/UX дизайна',
-                    'Умение создавать дизайны визуально привлекательных и функциональных интерфейсов'
-                ] 
-            },
-            { 
-                title: 'Личные качества', 
-                items: [
-                    'Креативность и творческий подход', 
-                    'Внимание к деталям',
-                    ''
-                ] 
-            },
-            { 
-                title: 'Работа в команде', 
-                items: [
-                    'Коммуникабельность и умение взаимодействовать с другими членами команды',  
-                    'Открытость к конструктивной критике и готовность к совместному улучшению проектов'
-                ] 
-            },
-            { 
-                title: 'Инициативность', 
-                items: [
-                    'Стремление к постоянному обучению и освоению новых инструментов и технологий', 
-                    'Заинтересованность в новейших трендах в области веб-дизайна'
-                ] 
-            }
-        ]
-    },
-    mobiledesigner: {
-        title: 'Mobile Designer📱',
-        description: 
-        `В роли стажера-программиста iOS у вас будет возможность работать над передовыми мобильными приложениями Apple, 
-        совершенствуя свои навыки вместе с опытными экспертами. Эта роль призвана предоставить вам глубокое понимание 
-        и практические навыки, которые станут трамплином для успешной карьеры в технологии. Вы будете погружены в среду, насыщенную обучением, создавая элегантные решения и 
-        расширяя границы возможностей на iOS. Если вы стремитесь начать путь, который объединяет инновации и образование, это начало вашего пути.`,
-        expectations: [
-            { 
-                title: 'Мобильная разработка', 
-                items: [
-                    'Понимание основ разработки для мобильных платформ', 
-                    'Знание Swift и SwiftUI'
-                ] 
-            },
-            { 
-                title: 'Работа с инструментами', 
-                items: [
-                    'Желательно умение работать с Apple Health Kit & iCloud SDK',
-                    'Умение работать с системой контроля версий Git'
-                ] 
-            },
-            { 
-                title: 'Инициативность и обучение', 
-                items: [
-                    'Готовность к обучению новым технологиям и методам разработки', 
-                    'Интерес к текущим трендам и лучшим практикам в мобильной разработке'
-                ] 
-            },
-            { 
-                title: 'Работа в команде', 
-                items: [
-                    'Способность эффективно работать в команде и взаимодействовать с другими разработчиками', 
-                    'Открытость к конструктивной критике и желание развиваться вместе с командой'
-                ] 
-            }
-        ]
-    },
-    frontend: {
-        title: 'Frontend Developer👨‍💻',
-        description: 
-        `Ищем увлеченного стажера-программиста для работы в нашей команде!
-        Ваша задача — создание красивых и функциональных пользовательских интерфейсов с использованием ReactJS, 
-        TypeScript и Bootstrap. Мы ищем талантливого специалиста, который сможет внести свой вклад в 
-        фронтенд-разработку нашего сайта.`,
-        expectations: [
-           {
-            title: 'Стек, работа с инструментами',
-            items: [
-                'Опыт разработки с использованием React и TypeScript', 
-                'Знание и опыт работы с Webpack',
-                'Умение работать с Git',
-
-            ] 
-           },
-           {
-            title: 'Веб-разработка',
-            items: [
-                'Понимание принципов атомарного дизайна и следование им', 
-                'Опыт работы с препроцессорами CSS',
-                'Адаптивная и отзывчивая верстка',
-            ] 
-           },
-           { 
-            title: 'Инициативность', 
-            items: [
-                'Готовность к изучению новых технологий и методов разработки', 
-                'Интерес к текущим трендам в веб-разработке и желание применить их в проектах'
-            ] 
-            },
-            { 
-                title: 'Работа в команде', 
-                items: [
-                    'Способность эффективно взаимодействовать с другими разработчиками, дизайнерами и менеджерами проектов', 
-                    'Открытость к конструктивной критике и способность работать над улучшением продукта вместе с командой'
-                ] 
-            }
-        ]
-    },
-    qa: {
-        title: 'Quality Assurance💯',
-        description: 
-        `Начните свою карьеру стажером по качеству программного обеспечения (QA), погружаясь в сложный мир тестирования веб- или мобильных приложений для Apple.
-        Эта возможность не ограничивается поиском ошибок; это практическое обучение, которое даст вам знания и навыки для успешного будущего в технологии. 
-        Ваш опыт здесь станет не просто стажировкой, а основой вашего пути к становлению экспертом в индустрии. Присоединяйтесь к нашей команде и 
-        превратите каждое тестирование в шаг к вашей мечтательной карьере.`,
-        expectations: [
-            { 
-                title: 'Тестирование', 
-                items: [
-                    'Умение разрабатывать тестовые сценарии и тест-кейсы, включая создание и выполнение тестов, а также анализ результатов', 
-                    'Понимание базовых принципов и методов тестирования программного обеспечения, включая тестирование функциональности, производительности и безопасности'
-                ] 
-            },
-            { 
-                title: 'Личные качества', 
-                items: [
-                    'Аналитическое мышление', 
-                    'Внимание к деталям'
-                ] 
-            },
-            { 
-                title: 'Инициативность', 
-                items: [
-                    'Желание учиться и развиваться', 
-                    'Готовность решать поставленные задачи и добиваться результатов'
-                ] 
-            },
-            { 
-                title: 'Документирование дефектов', 
-                items: [
-                    'Умение создать тикет, в котором будет указан ожидаемый и фактический результат, а также шаги для воспроизведения ошибки'
-                ] 
-            }
-        ]
-    },
-    ba: {
-        title: 'Business Analyst📊',
-        description: `Начните свою карьеру в роли аналитика, где вы сможете использовать данные для формирования будущего мобильных и веб-приложений.
-        Погружайтесь в тренды, выявляйте ключевые статистические данные и поддерживайте принятие решений 
-        на основе данных для улучшения нашего программного обеспечения. Ваши идеи будут определять наши следующие важные функции 
-        и оптимизировать взаимодействие в социальных сетях. Эта роль станет вашим трамплином к успешной карьере, предоставляя уникальные возможности для роста в постоянно развивающемся цифровом ландшафте. 
-        Поднимите свои навыки на новый уровень и помогите нам совершенствовать наши решения — вместе мы будем лидировать в мире технологий!`,
-expectations: [
-    { 
-        title: 'Анализ требований', 
-        items: [
-            'Сбор и документирование требований от заинтересованных сторон',
-            'Преобразование бизнес-целей в технические спецификации',
-            'Анализ текущих бизнес-процессов и выявление потребностей для их улучшения'
-        ] 
-    },
-    { 
-        title: 'Моделирование бизнес-процессов', 
-        items: [
-            'Использование методов моделирования бизнес-процессов (BPMN, UML)',
-            'Создание визуальных схем и диаграмм для упрощения понимания процессов',
-            'Выявление узких мест и предложений по оптимизации'
-        ] 
-    },
-    { 
-        title: 'Анализ данных', 
-        items: [
-            'Исследование рыночных тенденций и потребностей целевой аудитории',
-            'Анализ конкурентных преимуществ нашего продукта',
-            'Работа с инструментами анализа данных и отчетности (например, Excel, SQL)'
-        ] 
-    },
-    { 
-        title: 'Коммуникация и взаимодействие', 
-        items: [
-            'Эффективное взаимодействие с заказчиками и членами команды',
-            'Четкое формулирование требований и рекомендаций',
-            'Навыки активного слушания и умение задавать уточняющие вопросы'
-        ] 
-    },
-    { 
-        title: 'Критическое мышление и решение проблем', 
-        items: [
-            'Анализ и оценка информации для выявления проблем и возможностей',
-            'Разработка и проверка гипотез, предлагая обоснованные решения',
-            'Способность к критическому мышлению и принятию обоснованных решений'
-        ] 
-    },
-    { 
-        title: 'Желание учиться и развиваться', 
-        items: [
-            'Готовность к обучению новым методам и инструментам бизнес-анализа',
-            'Интерес к современным трендам и лучшим практикам в бизнес-анализе',
-            'Активное участие в профессиональном развитии и повышении квалификации'
-        ] 
-    }
-]
-    },
-    copywriter: {
-        title: 'Copywriter📝',
-        description: `Начните свою карьеру в роли аналитика, где вы сможете использовать данные для формирования будущего мобильных и веб-приложений.
-        Погружайтесь в тренды, выявляйте ключевые статистические данные и поддерживайте принятие решений на основе данных для улучшения 
-        нашего программного обеспечения. Ваши идеи будут определять наши следующие важные функции и оптимизировать взаимодействие в социальных сетях. 
-        Эта роль станет вашим трамплином к успешной карьере, предоставляя уникальные возможности для роста в постоянно развивающемся цифровом ландшафте. 
-        Поднимите свои навыки на новый уровень и помогите нам совершенствовать наши решения — вместе мы будем лидировать в мире технологий!`,
-        expectations: [
-            { 
-                title: 'Написание текстов', 
-                items: [
-                    'Для сайта', 
-                    'Для мобильных приложений'
-                ] 
-            },
-            { 
-                title: 'Ассоциативное мышление', 
-                items: [
-                    'Творческий подход к генерации идей', 
-                    'Желание учиться создавать цепляющие заголовки и тексты'
-                ] 
-            },
-            { 
-                title: 'Инициативность', 
-                items: [
-                    'Активное участие в обсуждениях', 
-                    'Предложение новых решений и вариантов'
-                ] 
-            },
-            { 
-                title: 'Личные качества', 
-                items: [
-                    'Желание получить опыт в реальном проекте', 
-                    'Высокий уровень эмпатии, достаточный для работы в международной команде'
-                ] 
-            },
-            { 
-                title: 'Работа с AI', 
-                items: [
-                    'Готовность обучаться и работать с инструментами ИИ для написания текстов'
-                ] 
-            },
-            { 
-                title: 'Ответственность', 
-                items: [
-                    'Готовность выполнять задачи, за которые беретесь', 
-                    'Умение работать самостоятельно'
-                ] 
-            }
-        ]
-    }
-};
-
-const internshipConditions = [
-    "— Это практика на безвозмездной основе на 2-3 месяца, — ваш труд не оплачивается, вы получаете знания и опыт работы",
-    "— Строгих часов нет, можно трудиться в любое, удобное для вас время",
-    "— Работа из дома на своем устройстве",
-    "— Если вам нужно будет отлучиться на несколько дней от проекта — это нормально, но просьба заранее предупредить чтобы ваша практика не закончилась заранее"
-] 
-
-const featuresOfWork = [
-    "— Дружелюбная и вдохновляющая атмосфера стартапа",
-    "— Уникальная возможность развиваться и становиться ключевым игроком в нашей команде",
-    "— Возможность влиять на здоровый образ жизни и благополучие наших пользователей"
-]
-
-const VacancyDetails: React.FC<VacancyDetailsProps> = ({ position }) => {
-    const data = vacancyData[position];
-
-    if (!data) {
-        return <p>Вакансия не найдена</p>;
-    }
-
-    return (
-        <>
-            <NavigationBar />
-            <div className="ms-base-page ms-base-new vacancy-details">
-                <section>
-                    <div className="row">
-                        <div className="col-lg-8">
-                            <h1>{data.title}</h1>
-                            <p className="p" style={{ fontWeight: 'bold' }}>Не оплачивается</p>
-                            <p style={{ color: '#212529' }}>Стажировка, удаленная работа</p>
-                        </div>
-                        <p style={{ color: '#212529', marginBottom: '0' }}>
-                            {data.description}
-                        </p>
-                    </div>
-                </section>
-
-                <section className='row'>
-                    <h2>Ожидания от кандидата</h2>
-                    <div className="row row-no-padding col-12 grid">
-                        {data.expectations.map((expectation: any, index: number) => (
-                            <div className="col-lg-6 mb-4" key={index} style={{justifyContent: 'flex-start'}}>
-                                <h3 className="mb-3">{expectation.title}</h3>
-                                <BulletList
-                                    items={expectation.items}
-                                    iconColor="#FFB340"
-                                    className="col-bullet-list"
-                                />
-                            </div>
-                        ))}
-                    </div>
-                </section>
-
-                <section className='row'>
-                    <h2>Условия стажировки</h2>
-                    <ul>
-                        {internshipConditions.map((condition: string, index: number) => (
-                            <li key={index}>{condition}</li>
-                        ))}
-                    </ul>
-                </section>
-
-                <section className='row'>
-                    <h2>Особенности</h2>
-                    <h3>Почему стоит выбрать нас:</h3>
-                    <ul>
-                        {featuresOfWork.map((feature: string, index: number) => (
-                            <li key={index}>{feature}</li>
-                        ))}
-                    </ul>
-                </section>
-
-                <section>
-                    <div className='row'>
-                        <h2>Отклик на стажировку</h2>
-                        <p>
-                            1. Заполните Google форму и отправьте ее нам<br/>
-                            2. Мы все внимательно изучим и обязательно вернемся к вам с обратной связью в кратчайшие сроки
-                        </p>
-                        <div>
-                            <button className='mt-2 ms-btn-large'>Откликнуться</button>
-                        </div>
-                    </div>
-                </section>
-            </div>
-        </>
-    );
-};
-
-export default VacancyDetails;
