+++ src/sass/molecule/bullet-list.scss
@@ -31,12 +31,12 @@
 }
 
 .col-bullet-list {
-  display: block; 
+  display: grid; 
+  gap: 1rem;
 
   li {
     display: flex; 
     align-items: flex-start; 
-    margin-bottom: 1rem;
 
     .bullet-list__icon {
       margin-top: 4px;
