+++ src/sass/page/vacancy-details.scss
@@ -1,12 +1,9 @@
 .vacancy-details {
-    .grid {
-        display: grid ;
-        width: 100% ;
-        grid-template-columns: repeat(2, 1fr);
-        gap: 1.5rem ;
-    }
-    
     #no-padding{
         padding: 0;
+    }
+
+    .h3 {
+        margin-bottom: 1.375rem;
     }
   }
